scamshot/requirements.txt
streamlit==1.28.0
fastapi==0.104.1
uvicorn==0.24.0.post1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
tldextract==5.1.1
nltk==3.8.1
textstat==0.7.3
clamd==1.0.2
pandas==2.1.1
scikit-learn==1.3.0
transformers==4.34.0
redis==4.6.0
python-multipart==0.0.6